# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: National Poets Evaluation Service API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List all submissions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, under_review, jury_assigned, evaluated, approved, rejected, archived]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionListResponse"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a poem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get submission by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
    patch:
      operationId: updateSubmission
      tags: [submissions]
      summary: Update submission status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubmissionRequest"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/assign-jury:
    post:
      operationId: assignJury
      tags: [submissions]
      summary: Assign jury members to a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignJuryRequest"
      responses:
        "200":
          description: Jury assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, reviewer, jury, poet]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
...
  /evaluations:
    get:
      operationId: listEvaluations
      tags: [evaluations]
      summary: List evaluations
      parameters:
        - name: submissionId
          in: query
          schema:
            type: integer
        - name: juryMemberId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of evaluations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationListResponse"
    post:
      operationId: createEvaluation
      tags: [evaluations]
      summary: Submit evaluation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvaluationRequest"
      responses:
        "201":
          description: Evaluation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, reviewer, jury, poet]
        status:
          type: string
          enum: [active, inactive]
        phone:
          type: string
        nationality:
          type: string
        bio:
          type: string
        bioAr:
          type: string
        createdAt:
          type: string
          format: date-time
        avatarUrl:
          type: string
      required: [id, name, email, role, status, createdAt]
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [users, total, page, limit]
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, reviewer, jury, poet]
        phone:
          type: string
        nationality:
          type: string
        bio:
          type: string
        bioAr:
          type: string
      required: [name, email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, reviewer, jury, poet]
        status:
          type: string
          enum: [active, inactive]
        phone:
          type: string
        nationality:
          type: string
        bio:
          type: string
        bioAr:
          type: string
    Submission:
      type: object
      properties:
        id:
          type: integer
        referenceNumber:
          type: string
        poetName:
          type: string
        poetNameAr:
          type: string
        poetEmail:
          type: string
        poetPhone:
          type: string
        poetNationality:
          type: string
        poemTitle:
          type: string
        poemTitleAr:
          type: string
        poemContent:
          type: string
        poemType:
          type: string
          enum: [nabati, classical, modern]
        status:
          type: string
          enum: [pending, under_review, jury_assigned, evaluated, approved, rejected, archived]
        submittedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        reviewerNotes:
          type: string
        finalDecision:
          type: string
        finalScore:
          type: number
        assignedJuryIds:
          type: array
          items:
            type: integer
        evaluations:
          type: array
          items:
            $ref: "#/components/schemas/Evaluation"
      required: [id, referenceNumber, poetName, poetEmail, poemTitle, poemContent, poemType, status, submittedAt]
    SubmissionListResponse:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/Submission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [submissions, total, page, limit]
    CreateSubmissionRequest:
      type: object
      properties:
        poetName:
          type: string
        poetNameAr:
          type: string
        poetEmail:
          type: string
        poetPhone:
          type: string
        poetNationality:
          type: string
        poemTitle:
          type: string
        poemTitleAr:
          type: string
        poemContent:
          type: string
        poemType:
          type: string
          enum: [nabati, classical, modern]
        competitionId:
          type: integer
      required: [poetName, poetEmail, poemTitle, poemContent, poemType]
    UpdateSubmissionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, under_review, jury_assigned, evaluated, approved, rejected, archived]
        reviewerNotes:
          type: string
        finalDecision:
          type: string
    AssignJuryRequest:
      type: object
      properties:
        juryMemberIds:
          type: array
          items:
            type: integer
      required: [juryMemberIds]
...
    Evaluation:
      type: object
      properties:
        id:
          type: integer
        submissionId:
          type: integer
        juryMemberId:
          type: integer
        juryMemberName:
          type: string
        linguisticScore:
          type: number
          minimum: 0
          maximum: 10
        poeticScore:
          type: number
          minimum: 0
          maximum: 10
        originalityScore:
          type: number
          minimum: 0
          maximum: 10
        emotionalScore:
          type: number
          minimum: 0
          maximum: 10
        culturalScore:
          type: number
          minimum: 0
          maximum: 10
        totalScore:
          type: number
        notes:
          type: string
        recommendation:
          type: string
          enum: [approve, reject, needs_revision]
        submittedAt:
          type: string
          format: date-time
      required: [id, submissionId, juryMemberId, totalScore, recommendation]
    EvaluationListResponse:
      type: object
      properties:
        evaluations:
          type: array
          items:
            $ref: "#/components/schemas/Evaluation"
        total:
          type: integer
      required: [evaluations, total]
    CreateEvaluationRequest:
      type: object
      properties:
        submissionId:
          type: integer
        juryMemberId:
          type: integer
        linguisticScore:
          type: number
        poeticScore:
          type: number
        originalityScore:
          type: number
        emotionalScore:
          type: number
        culturalScore:
          type: number
        notes:
          type: string
        recommendation:
          type: string
          enum: [approve, reject, needs_revision]
      required: [submissionId, juryMemberId, linguisticScore, poeticScore, originalityScore, emotionalScore, culturalScore, recommendation]
...