# lib/api-zod/src/generated/types/createEvaluationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * National Poets Evaluation Service API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEvaluationRequestRecommendation } from "./createEvaluationRequestRecommendation";

export interface CreateEvaluationRequest {
  submissionId: number;
  juryMemberId: number;
  linguisticScore: number;
  poeticScore: number;
  originalityScore: number;
  emotionalScore: number;
  culturalScore: number;
  notes?: string;
  recommendation: CreateEvaluationRequestRecommendation;
}