# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * National Poets Evaluation Service API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  name: string;
  nameAr?: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  phone?: string;
  nationality?: string;
  bio?: string;
  bioAr?: string;
  createdAt: Date;
  avatarUrl?: string;
}